fix(iso): Correct ESP size calculation and boot catalog entry
`fat::create_fat_image` now returns the logical size of the created FAT image in 512-byte sectors. `iso::builder::build_iso` uses this logical size to calculate the ESP size in ISO (2048-byte) sectors for the ISO builder's internal tracking.

When creating the UEFI boot catalog entry, `iso::builder_utils::create_uefi_esp_boot_entry` now correctly converts the ESP size from ISO sectors back to 512-byte sectors, as required by the boot catalog specification.

This ensures consistent and accurate handling of the EFI System Partition (ESP) size throughout the ISO creation process, resolving discrepancies between 512-byte and 2048-byte sector units and ensuring the boot catalog correctly reflects the ESP's logical size. The `build_iso` function now also returns the logical FAT image size in 512-byte sectors.
