+++ src/iso/builder.rs
@@ -263,11 +263,12 @@ pub fn build_iso(
     iso_path: &Path,
     image: &IsoImage,
     is_isohybrid: bool,
-) -> io::Result<(PathBuf, Option<NamedTempFile>, File)> {
+) -> io::Result<(PathBuf, Option<NamedTempFile>, File, Option<u32>)> { // Added Option<u32> for logical_fat_size_512_sectors
     let mut iso_builder = IsoBuilder::new();
     iso_builder.set_isohybrid(is_isohybrid);
 
     let mut temp_fat_file_holder: Option<NamedTempFile> = None;
+    let mut logical_fat_size_512_sectors: Option<u32> = None; // Declare here
 
     // Create the ISO file
     let mut iso_file = File::create(iso_path)?;
@@ -280,15 +281,17 @@ pub fn build_iso(
             let path = temp_fat_file.path().to_path_buf();
             temp_fat_file_holder = Some(temp_fat_file);
 
-            fat::create_fat_image(&path, &uefi_boot.boot_image, &uefi_boot.kernel_image)?;
+            let size_512_sectors =
+                fat::create_fat_image(&path, &uefi_boot.boot_image, &uefi_boot.kernel_image)?;
+            logical_fat_size_512_sectors = Some(size_512_sectors); // Assign here
 
-            let fat_img_metadata = std::fs::metadata(&path)?;
-            let calculated_esp_size_sectors =
-                (fat_img_metadata.len() as u32).div_ceil(crate::utils::ISO_SECTOR_SIZE as u32);
+            // Convert logical FAT size from 512-byte sectors to ISO 2048-byte sectors
+            let calculated_esp_size_iso_sectors =
+                size_512_sectors.div_ceil(4); // 1 ISO sector = 4 * 512-byte sectors
 
             // Store ESP LBA and size for the boot catalog
             iso_builder.esp_lba = Some(ESP_START_LBA); // ESP starts at LBA 34 for hybrid ISOs
-            iso_builder.esp_size_sectors = Some(calculated_esp_size_sectors);
+            iso_builder.esp_size_sectors = Some(calculated_esp_size_iso_sectors);
 
             // Copy the FAT image to the ISO file at the designated ESP LBA (34)
             iso_file.seek(SeekFrom::Start(
@@ -326,7 +329,12 @@ pub fn build_iso(
     // The iso_file is already the final_iso_file
     let final_iso_file = iso_file;
 
-    Ok((iso_path.to_path_buf(), temp_fat_file_holder, final_iso_file))
+    Ok((
+        iso_path.to_path_buf(),
+        temp_fat_file_holder,
+        final_iso_file,
+        logical_fat_size_512_sectors, // Return this value
+    ))
 }
 
 #[cfg(test)]
