+++ tests/integration.rs
@@ -1,447 +1,1 @@
-// tests/integration.rs
-use std::{
-    fs::File,
-    io::{self, Error, ErrorKind, Read, Seek, SeekFrom},
-    path::{Path, PathBuf},
-    process::Command,
-};
-
-use isobemak::{BootInfo, IsoImage, IsoImageFile, UefiBootInfo, build_iso};
-use tempfile::tempdir;
-
-fn run_command(command: &str, args: &[&str]) -> io::Result<String> {
-    let output = Command::new(command).args(args).output()?;
-
-    if output.status.success() {
-        Ok(String::from_utf8_lossy(&output.stdout).into_owned())
-    } else {
-        Err(Error::new(
-            ErrorKind::Other,
-            format!(
-                "Command `{}` failed with exit code {:?}\nStdout: {}\nStderr: {}",
-                command,
-                output.status.code(),
-                String::from_utf8_lossy(&output.stdout),
-                String::from_utf8_lossy(&output.stderr)
-            ),
-        ))
-    }
-}
-fn setup_integration_test_files(temp_dir: &Path) -> io::Result<(PathBuf, PathBuf, PathBuf)> {
-    // Create dummy files needed for the ISO image
-    let bootx64_path = temp_dir.join("bootx64.efi");
-    std::fs::write(&bootx64_path, vec![0u8; 64 * 1024])?;
-
-    let kernel_path = temp_dir.join("kernel.elf");
-    std::fs::write(&kernel_path, vec![0u8; 16 * 1024])?;
-
-    let iso_path = temp_dir.join("test.iso");
-
-    Ok((bootx64_path, kernel_path, iso_path))
-}
-
-#[test]
-fn test_create_disk_and_iso() -> io::Result<()> {
-    let temp_dir = tempdir()?;
-    let temp_dir_path = temp_dir.path();
-    println!("Temp dir: {:?}", &temp_dir_path);
-
-    // Setup files and paths
-    let (bootx64_path, kernel_path, iso_path) = setup_integration_test_files(&temp_dir_path)?;
-
-    let iso_image = IsoImage {
-        files: vec![
-            IsoImageFile {
-                source: bootx64_path.clone(),
-                destination: "EFI/BOOT/BOOTX64.EFI".to_string(),
-            },
-            IsoImageFile {
-                source: kernel_path.clone(),
-                destination: "EFI/BOOT/KERNEL.EFI".to_string(),
-            },
-        ],
-        boot_info: BootInfo {
-            bios_boot: None, // Not testing BIOS boot in this specific test
-            uefi_boot: Some(UefiBootInfo {
-                boot_image: bootx64_path.clone(),
-                kernel_image: kernel_path.clone(),
-                destination_in_iso: "EFI/BOOT/BOOTX64.EFI".to_string(),
-            }),
-        },
-    };
-
-    // Call the main function with correct arguments
-    build_iso(&iso_path, &iso_image, false)?;
-    // Assert that the ISO file was created
-    assert!(iso_path.exists());
-
-    // Verify ISO content using isoinfo
-    let isoinfo_d_output = run_command("isoinfo", &["-d", "-i", iso_path.to_str().unwrap()])?;
-    println!("isoinfo -d output:\n{}", isoinfo_d_output);
-    assert!(isoinfo_d_output.contains("Volume id: ISOBEMAKI"));
-
-    // Verify Nsect value in the boot catalog
-    let mut iso_file_for_nsect_check = File::open(&iso_path)?;
-    let boot_catalog_start_pos = isobemak::iso::boot_catalog::LBA_BOOT_CATALOG as u64
-        * isobemak::utils::ISO_SECTOR_SIZE as u64;
-    iso_file_for_nsect_check.seek(SeekFrom::Start(boot_catalog_start_pos))?;
-
-    let mut boot_catalog_sector = [0u8; isobemak::utils::ISO_SECTOR_SIZE];
-    iso_file_for_nsect_check.read_exact(&mut boot_catalog_sector)?;
-
-    // The first boot entry starts after the 32-byte validation entry.
-    // Nsect is at offset 6-7 within the boot entry.
-    let boot_entry_offset = 32; // Validation entry size
-    let nsect_offset_in_entry = 6;
-    let nsect_bytes_start = boot_entry_offset + nsect_offset_in_entry;
-
-    let nsect = u16::from_le_bytes([
-        boot_catalog_sector[nsect_bytes_start],
-        boot_catalog_sector[nsect_bytes_start + 1],
-    ]);
-
-    let efi_size = 64 * 1024; // BOOTX64.EFI size from setup_integration_test_files
-    const EL_TORITO_SECTOR_SIZE: u64 = 512;
-    let expected_sectors = (efi_size as u64).div_ceil(EL_TORITO_SECTOR_SIZE) as u16;
-
-    assert_eq!(
-        nsect, expected_sectors,
-        "Nsect value in boot catalog is incorrect"
-    );
-    println!("Verified Nsect: {} (expected: {})", nsect, expected_sectors);
-
-    let isoinfo_l_output = run_command("isoinfo", &["-l", "-i", iso_path.to_str().unwrap()])?;
-    println!("isoinfo -l output:\n{}", isoinfo_l_output);
-    assert!(isoinfo_l_output.contains("BOOTX64.EFI;1"));
-    assert!(isoinfo_l_output.contains("KERNEL.EFI;1"));
-
-    // Verify ISO content using 7z
-    let sevenz_output = Command::new("7z")
-        .args(&["l", iso_path.to_str().unwrap()])
-        .output()?;
-    let sevenz_l_output = String::from_utf8_lossy(&sevenz_output.stdout).into_owned();
-    println!("7z l output:\n{}", sevenz_l_output);
-    assert!(sevenz_l_output.contains("EFI/BOOT/BOOTX64.EFI"));
-    assert!(sevenz_l_output.contains("EFI/BOOT/KERNEL.EFI"));
-
-    // Extract the UEFI boot image and verify with dumpet
-    let extract_dir = temp_dir_path.join("extracted");
-    std::fs::create_dir_all(&extract_dir)?;
-    let _extract_output = Command::new("7z")
-        .args(&[
-            "x",
-            iso_path.to_str().unwrap(),
-            "-o",
-            extract_dir.to_str().unwrap(),
-        ])
-        .output()?;
-    // Proceed even if there are warnings
-
-    let extracted_bootx64_path = extract_dir.join("EFI/BOOT/BOOTX64.EFI");
-    // Skip extraction assertion due to 7z warning, but verify file size if exists
-    if extracted_bootx64_path.exists() {
-        let dumpet_output = run_command("dumpet", &[extracted_bootx64_path.to_str().unwrap()])?;
-        println!("dumpet output:\n{}", dumpet_output);
-        assert!(dumpet_output.contains("EFI boot image"));
-    } else {
-        println!("Extraction failed, but listing succeeded");
-    }
-
-    // Verify the boot catalog validation entry checksum
-    let mut iso_file = File::open(iso_path)?;
-    iso_file.seek(SeekFrom::Start(
-        isobemak::iso::boot_catalog::LBA_BOOT_CATALOG as u64 * 2048,
-    ))?;
-    let mut boot_catalog = [0u8; 32]; // Only need the validation entry
-    iso_file.read_exact(&mut boot_catalog)?;
-
-    let mut sum: u16 = 0;
-    for chunk in boot_catalog.chunks_exact(2) {
-        sum = sum.wrapping_add(u16::from_le_bytes(chunk.try_into().unwrap()));
-    }
-
-    assert_eq!(sum, 0, "Boot catalog validation entry checksum should be 0");
-
-    // Perform deeper binary verification of ISO structures
-    verify_iso_binary_structures(&mut iso_file)?;
-
-    Ok(())
-}
-
-#[test]
-fn test_create_isohybrid_uefi_iso() -> io::Result<()> {
-    let temp_dir = tempdir()?;
-    let temp_dir_path = temp_dir.path();
-    println!("Temp dir for isohybrid UEFI test: {:?}", &temp_dir_path);
-
-    // Setup files and paths
-    let (bootx64_path, kernel_path, iso_path) = setup_integration_test_files(&temp_dir_path)?;
-
-    let iso_image = IsoImage {
-        files: vec![
-            IsoImageFile {
-                source: bootx64_path.clone(),
-                destination: "EFI/BOOT/BOOTX64.EFI".to_string(),
-            },
-            IsoImageFile {
-                source: kernel_path.clone(),
-                destination: "EFI/BOOT/KERNEL.EFI".to_string(),
-            },
-        ],
-        boot_info: BootInfo {
-            bios_boot: None,
-            uefi_boot: Some(UefiBootInfo {
-                boot_image: bootx64_path.clone(),
-                kernel_image: kernel_path.clone(),
-                destination_in_iso: "EFI/BOOT/BOOTX64.EFI".to_string(),
-            }),
-        },
-    };
-
-    // Call the main function with is_isohybrid set to true
-    let (fat_image_path, _temp_fat_file_holder, _iso_file) =
-        build_iso(&iso_path, &iso_image, true)?;
-    assert!(iso_path.exists());
-    assert!(fat_image_path.exists());
-    // _iso_file is kept in scope to ensure the ISO file remains open for verification
-    // _temp_fat_file_holder is kept in scope to prevent the temporary file from being deleted
-
-    // Verify ISO content using isoinfo -d
-    let isoinfo_d_output = run_command("isoinfo", &["-d", "-i", iso_path.to_str().unwrap()])?;
-    println!("isoinfo -d output (isohybrid):\n{}", isoinfo_d_output);
-    assert!(isoinfo_d_output.contains("Volume id: ISOBEMAKI"));
-
-    // Verify the UEFI boot catalog entry
-    let mut iso_file_for_nsect_check = File::open(&iso_path)?;
-    let boot_catalog_start_pos = isobemak::iso::boot_catalog::LBA_BOOT_CATALOG as u64
-        * isobemak::utils::ISO_SECTOR_SIZE as u64;
-    iso_file_for_nsect_check.seek(SeekFrom::Start(boot_catalog_start_pos))?;
-
-    let mut boot_catalog_sector = [0u8; isobemak::utils::ISO_SECTOR_SIZE];
-    iso_file_for_nsect_check.read_exact(&mut boot_catalog_sector)?;
-
-    // Verify Validation Entry's platform ID
-    assert_eq!(
-        boot_catalog_sector[1],
-        isobemak::iso::boot_catalog::BOOT_CATALOG_EFI_PLATFORM_ID,
-        "Validation entry platform ID is not EFI"
-    );
-
-    // Verify the first Boot Entry (which should be the UEFI one in this test case)
-    let boot_entry_offset = 32; // After the 32-byte validation entry
-    let boot_entry_bytes = &boot_catalog_sector[boot_entry_offset..boot_entry_offset + 32];
-
-    let uefi_boot_indicator = boot_entry_bytes[0];
-    let uefi_boot_sectors = u16::from_le_bytes(boot_entry_bytes[6..8].try_into().unwrap());
-    let uefi_boot_lba = u32::from_le_bytes(boot_entry_bytes[8..12].try_into().unwrap());
-
-    assert_eq!(
-        uefi_boot_indicator,
-        isobemak::iso::boot_catalog::BOOT_CATALOG_BOOT_ENTRY_HEADER_ID,
-        "UEFI boot entry is not marked bootable"
-    );
-    assert_eq!(
-        uefi_boot_lba,
-        isobemak::ESP_START_LBA,
-        "UEFI boot LBA in boot catalog is incorrect"
-    );
-
-    let fat_img_metadata = std::fs::metadata(&fat_image_path)?;
-    let expected_esp_sectors =
-        (fat_img_metadata.len() as u32).div_ceil(isobemak::utils::ISO_SECTOR_SIZE as u32) as u16;
-
-    assert_eq!(
-        uefi_boot_sectors, expected_esp_sectors,
-        "UEFI boot sectors in boot catalog is incorrect"
-    );
-    println!(
-        "Verified UEFI boot entry: LBA={}, Sectors={} (expected: {})",
-        uefi_boot_lba, uefi_boot_sectors, expected_esp_sectors
-    );
-
-    // Verify ISO content using isoinfo -l
-    let isoinfo_l_output = run_command("isoinfo", &["-l", "-i", iso_path.to_str().unwrap()])?;
-    println!("isoinfo -l output (isohybrid):\n{}", isoinfo_l_output);
-    assert!(isoinfo_l_output.contains("BOOTX64.EFI;1"));
-    assert!(isoinfo_l_output.contains("KERNEL.EFI;1"));
-
-    // Verify the boot catalog validation entry checksum
-    let mut iso_file = File::open(iso_path)?;
-    iso_file.seek(SeekFrom::Start(
-        isobemak::iso::boot_catalog::LBA_BOOT_CATALOG as u64 * 2048,
-    ))?;
-    let mut boot_catalog = [0u8; 32]; // Only need the validation entry
-    iso_file.read_exact(&mut boot_catalog)?;
-
-    let mut sum: u16 = 0;
-    for chunk in boot_catalog.chunks_exact(2) {
-        sum = sum.wrapping_add(u16::from_le_bytes(chunk.try_into().unwrap()));
-    }
-
-    assert_eq!(sum, 0, "Boot catalog validation entry checksum should be 0");
-
-    // Perform deeper binary verification of ISO structures
-    verify_iso_binary_structures(&mut iso_file)?;
-
-    Ok(())
-}
-
-#[test]
-fn test_iso_integrity_and_boot_modes() -> io::Result<()> {
-    let temp_dir = tempdir()?;
-    let temp_dir_path = temp_dir.path();
-    println!("Temp dir for integrity test: {:?}", &temp_dir_path);
-
-    // Setup files and paths for an ISO with both BIOS and UEFI boot
-    let bios_boot_image_path = temp_dir_path.join("isolinux.bin");
-    let mut bios_boot_image = vec![0u8; 512];
-    bios_boot_image[510..512].copy_from_slice(&0xAA55u16.to_le_bytes());
-    std::fs::write(&bios_boot_image_path, bios_boot_image)?; // A dummy 512-byte boot image with signature
-    let bios_cfg_path = temp_dir_path.join("isolinux.cfg");
-    std::fs::write(&bios_cfg_path, b"default menu.c32")?;
-
-    let bootx64_path = temp_dir_path.join("bootx64.efi");
-    std::fs::write(&bootx64_path, vec![0u8; 64 * 1024])?;
-
-    let kernel_path = temp_dir_path.join("kernel.elf");
-    std::fs::write(&kernel_path, vec![0u8; 16 * 1024])?;
-
-    let iso_path = temp_dir_path.join("integrity_test.iso");
-
-    let iso_image = isobemak::IsoImage {
-        files: vec![
-            isobemak::IsoImageFile {
-                source: bios_cfg_path.clone(),
-                destination: "isolinux/isolinux.cfg".to_string(),
-            },
-            isobemak::IsoImageFile {
-                source: bootx64_path.clone(),
-                destination: "EFI/BOOT/BOOTX64.EFI".to_string(),
-            },
-            isobemak::IsoImageFile {
-                source: kernel_path.clone(),
-                destination: "EFI/BOOT/KERNEL.EFI".to_string(),
-            },
-        ],
-        boot_info: isobemak::BootInfo {
-            bios_boot: Some(isobemak::BiosBootInfo {
-                boot_catalog: PathBuf::from("BOOT.CAT"),
-                boot_image: bios_boot_image_path.clone(),
-                destination_in_iso: "isolinux/isolinux.bin".to_string(),
-            }),
-            uefi_boot: Some(isobemak::UefiBootInfo {
-                boot_image: bootx64_path.clone(),
-                kernel_image: kernel_path.clone(),
-                destination_in_iso: "EFI/BOOT/BOOTX64.EFI".to_string(),
-            }),
-        },
-    };
-
-    // Build the ISO
-    build_iso(&iso_path, &iso_image, true)?;
-    assert!(iso_path.exists());
-
-    // 1. Verify ISO integrity using md5sum
-    let md5sum_output = run_command("md5sum", &[iso_path.to_str().unwrap()])?;
-    println!("md5sum output:\n{}", md5sum_output);
-    assert!(md5sum_output.contains(iso_path.file_name().unwrap().to_str().unwrap()));
-    // A more robust check would be to compare against a known good checksum,
-    // but for a generated ISO, simply ensuring the command runs and produces output is a start.
-
-    // 2. Verify BIOS (El Torito) boot entry
-    let isoinfo_d_output = run_command("isoinfo", &["-d", "-i", iso_path.to_str().unwrap()])?;
-    println!("isoinfo -d output (integrity test):\n{}", isoinfo_d_output);
-    assert!(isoinfo_d_output.contains("El Torito VD version 1 found")); // Updated assertion
-    assert!(isoinfo_d_output.contains("Arch 0 (x86)")); // Updated assertion
-    assert!(isoinfo_d_output.contains("Boot media 0 (No Emulation Boot)")); // Updated assertion
-    // Removed assertion for "EFI boot entry is present" as isoinfo -d does not output this string directly.
-    // Detailed UEFI boot entry verification is handled in `test_create_isohybrid_uefi_iso`.
-
-    // Extract the BIOS boot image and check its signature (0xAA55)
-    // This requires knowing the LBA of the boot image from the boot catalog.
-    // For simplicity, we'll assume the first boot entry is the BIOS one and extract it.
-    // A more robust solution would parse the boot catalog directly.
-    let extract_dir = temp_dir_path.join("extracted_bios_boot");
-    std::fs::create_dir_all(&extract_dir)?;
-    run_command(
-        "7z",
-        &[
-            "x",
-            iso_path.to_str().unwrap(),
-            &format!("-o{}", extract_dir.to_str().unwrap()),
-            "isolinux/isolinux.bin", // Assuming this is the BIOS boot image
-        ],
-    )?;
-
-    let extracted_bios_boot_path = extract_dir.join("isolinux/isolinux.bin");
-    if extracted_bios_boot_path.exists() {
-        let mut boot_image_file = File::open(&extracted_bios_boot_path)?;
-        let mut boot_sector = [0u8; 512];
-        boot_image_file.read_exact(&mut boot_sector)?;
-        // Check for boot signature 0xAA55 at offset 510-511
-        assert_eq!(
-            u16::from_le_bytes([boot_sector[510], boot_sector[511]]),
-            0xAA55,
-            "BIOS boot image does not have the expected boot signature (0xAA55)"
-        );
-        println!("Verified BIOS boot image signature (0xAA55)");
-    } else {
-        println!(
-            "Warning: isolinux/isolinux.bin not extracted or found for BIOS boot signature check."
-        );
-    }
-
-    // 3. Verify UEFI boot entry
-    // The `test_create_isohybrid_uefi_iso` already performs detailed UEFI boot entry verification.
-    // Removed assertion for "EFI boot entry is present" as isoinfo -d does not output this string directly.
-
-    Ok(())
-}
-
-/// Verifies critical binary structures within the generated ISO file.
-fn verify_iso_binary_structures(iso_file: &mut File) -> io::Result<()> {
-    const ISO_SECTOR_SIZE: u64 = 2048;
-
-    // 1. Verify Primary Volume Descriptor (PVD) at LBA 16
-    iso_file.seek(SeekFrom::Start(16 * ISO_SECTOR_SIZE))?;
-    let mut pvd_header = [0u8; 6];
-    iso_file.read_exact(&mut pvd_header)?;
-    assert_eq!(
-        &pvd_header,
-        &[0x01, b'C', b'D', b'0', b'0', b'1'],
-        "PVD identifier 'CD001' not found at LBA 16"
-    );
-
-    // 2. Verify Boot Record Volume Descriptor (BRVD) at LBA 17
-    iso_file.seek(SeekFrom::Start(17 * ISO_SECTOR_SIZE))?;
-    let mut brvd_header = [0u8; 37];
-    iso_file.read_exact(&mut brvd_header)?;
-    assert_eq!(
-        &brvd_header[0..7],
-        &[0x00, b'C', b'D', b'0', b'0', b'1', 0x01],
-        "BRVD identifier 'CD001' not found at LBA 17"
-    );
-    assert_eq!(
-        &brvd_header[7..30],
-        b"EL TORITO SPECIFICATION",
-        "BRVD boot identifier 'EL TORITO SPECIFICATION' not found"
-    );
-
-    // 3. Re-verify the boot catalog validation entry checksum at LBA 19
-    iso_file.seek(SeekFrom::Start(
-        isobemak::iso::boot_catalog::LBA_BOOT_CATALOG as u64 * ISO_SECTOR_SIZE,
-    ))?;
-    let mut boot_catalog = [0u8; 32]; // Only need the validation entry
-    iso_file.read_exact(&mut boot_catalog)?;
-
-    let mut sum: u16 = 0;
-    for chunk in boot_catalog.chunks_exact(2) {
-        sum = sum.wrapping_add(u16::from_le_bytes(chunk.try_into().unwrap()));
-    }
-    assert_eq!(
-        sum, 0,
-        "Boot catalog validation entry checksum should be 0 (re-verification)"
-    );
-
-    Ok(())
-}
+mod integration_tests;
