+++ tests/integration_tests/mod.rs
@@ -1,0 +1,4 @@
+pub mod common;
+pub mod basic_iso;
+pub mod isohybrid_uefi;
+pub mod integrity_and_boot;
